+++ lizmap/plugin.py
@@ -782,10 +782,9 @@ class Lizmap:
         self.action.triggered.connect(self.run)
 
         # Open the online help
-        if Qgis.QGIS_VERSION_INT >= 31000:
-            self.help_action = QAction(icon, 'Lizmap', self.iface.mainWindow())
-            self.iface.pluginHelpMenu().addAction(self.help_action)
-            self.help_action.triggered.connect(self.show_help)
+        self.help_action = QAction(icon, 'Lizmap', self.iface.mainWindow())
+        self.iface.pluginHelpMenu().addAction(self.help_action)
+        self.help_action.triggered.connect(self.show_help)
 
         # connect Lizmap signals and functions
         self.dlg.buttonBox.button(QDialogButtonBox.Cancel).clicked.connect(self.dlg.close)
@@ -973,7 +972,7 @@ class Lizmap:
         self.iface.webMenu().removeAction(self.action)
         self.iface.removeWebToolBarIcon(self.action)
 
-        if Qgis.QGIS_VERSION_INT >= 31000 and self.help_action:
+        if self.help_action:
             self.iface.pluginHelpMenu().removeAction(self.help_action)
             del self.help_action
 
@@ -1005,8 +1004,7 @@ class Lizmap:
 
         if has_geom:
             qgis_popup.setFlags(qgis_popup.flags() | Qt.ItemIsEnabled)
-
-        if not has_geom and Qgis.QGIS_VERSION_INT >= 31000:
+        else:
             qgis_popup.setFlags(qgis_popup.flags() | Qt.ItemIsEnabled)
 
     def show_help(self):