+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Desktop - Always export the new checkbox for a fixed scale in the overview map
 * Server - Fix Python exception when GetFeatureInfo does not have a feature ID
 
 ## 3.7.7 - 2022-04-13