+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Always export the new checkbox for a fixed scale in the overview map
 * Remove the code related to the server side. It's a new plugin called "Lizmap server" now.
+* Raise the QGIS minimum version to 3.10
 
 ## 3.7.7 - 2022-04-13